+++ src/ResourceManager/Sql/ChangeLog.md
@@ -18,6 +18,12 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Add commandlets for new feature - Instance Failover Groups
+	- Get-AzureRmSqlDatabaseInstanceFailoverGroup get the Instance Failover Group entity
+	- New-AzureRmSqlDatabaseInstanceFailoverGroup creates a new Instance Failover Group
+	- Remove-AzureRmSqlDatabaseInstanceFailoverGroup Instance Failover Group deletes the Failover Group
+	- Set-AzureRmSqlDatabaseInstanceFailoverGroup set Azure Sql Database Failover Policy and Grace Period entities of the Instance Failover Group
+	- Switch-AzureRmSqlDatabaseInstanceFailoverGroup issues the failover operation with data loss or without data loss
 * Set minimum dependency of module to PowerShell 5.0
 
 ## Version 4.4.0